+++ XADMaster/XADRAR13CryptHandle.m
@@ -25,12 +25,12 @@ static inline uint16_t rol8(uint16_t val,int n) { return (val<<n)|(val>>(8-n)); 
 
 -(void)resetByteStream
 {
-	int passlength=[password length];
+	NSInteger passlength=[password length];
 	const uint8_t *passbytes=[password bytes];
 
 	key1=key2=key3=0;
 
-	for(int i=0;i<passlength;i++)
+	for(NSInteger i=0;i<passlength;i++)
 	{
 		uint8_t c=passbytes[i];
 		key1+=c;
