+++ XADMaster/XADSimpleUnarchiver.m
@@ -759,8 +759,8 @@
 	NSMutableDictionary *parent=renames;
 	NSString *path=unpackdestination;
 	NSArray *components=[safefilename pathComponents];
-	int numcomponents=[components count];
-	for(int i=0;i<numcomponents;i++)
+	NSInteger numcomponents=[components count];
+	for(NSInteger i=0;i<numcomponents;i++)
 	{
 		NSString *component=components[i];
 		NSMutableDictionary *pathdict=parent[component];
