+++ XADMaster/XADTarParser.m
@@ -458,7 +458,7 @@
 		[handle seekToFileOffset:offset];
 
 		// Check if there is a terminating null byte, and eliminate it.
-		int length=[longHeader length];
+		NSInteger length=[longHeader length];
 		const uint8_t *bytes=[longHeader bytes];
 		if(length>0 && bytes[length-1]==0) longHeader=[longHeader subdataWithRange:NSMakeRange(0,length-1)];
